+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added new interactive command `cider-eval-last-expression-to-repl` (`C-c M-e`). The command will output the result
 of the evaluated code to the REPL buffer, so you can easily play with the output there afterwords.
 * Added new interactive command `cider-insert-last-expression-in-repl` (`C-c M-p`).
+* Added new interactive command `cider-eval-last-expression-and-replace` (`C-c C-w`).
 
 ### Changes